Snap.Snaplet.MongoDB.Functions.S

Documentation

eitherWithDB' :: HasMongoDB' app m => AccessMode -> Action IO a -> m (Either Failure a)

Database access function.

argument: AccessMode.

argument: Action to perform.

Returns: Either Failure or the action's result.

Example:

 eitherWithDB' UnconfirmedWrites $ insert "test-collection" ["some_field" = "something" ]

eitherWithDB :: HasMongoDB' app m => Action IO a -> m (Either Failure a)

Database access function.

argument: Action to perform. (Defaults to UnconfirmedWrites AccessMode)

Returns: Either Failure or the action's result.

Example:

 eitherWithDB $ insert "test-collection" ["some_field" = "something" ]

maybeWithDB :: HasMongoDB' app m => Action IO a -> m (Maybe a)

Database access function.

argument: Action to perform. (Defaults to UnconfirmedWrites AccessMode)

Returns: Nothing in case of failure or Just the rsult of the action.

Example:

 maybeWithDB $ insert "test-collection" ["some_field" = "something" ]

maybeWithDB' :: HasMongoDB' app m => AccessMode -> Action IO a -> m (Maybe a)

Database access function.

argument: AccessMode.

argument: Action to perform.

Returns: Nothing in case of failure or Just the rsult of the action.

Example:

 maybeWithDB' UnconfirmedWrites $ insert "test-collection" ["some_field" = "something" ]

unsafeWithDB :: HasMongoDB' app m => Action IO a -> m a

Database access function.

argument: Action to perform. (Defaults to UnconfirmedWrites AccessMode)

Returns: The action's result; in case of failure error is called.

Example:

 unsafeWithDB $ insert "test-collection" ["some_field" = "something" ]

unsafeWithDB' :: HasMongoDB' app m => AccessMode -> Action IO a -> m a

Database access function.

argument: AccessMode.

argument: Action to perform.

Returns: The action's result; in case of failure error is called.

Example:

 unsafeWithDB' UnconfirmedWrites $ insert "test-collection" ["some_field" = "something" ]